Generate grpc protobufs using the protoc-gen-go-grpc plugin
Previously grpc protobufs was generated by the protoc-gen-go plugin by
passing the flag `plugins=grpc`. In Carousell we use protoc-gen-go
provided by the module `github.com/golang/protobuf`. Unfortunately this
module contains only a limited functionality subset of the
protoc-gen-go-grpc plugin.

In particular, the later versions of protoc-gen-go-grpc export the
grpc.ServiceDesc variable (renamed from `_ServiceName_serviceDesc` to
`ServiceName_ServiceDesc`. If we continue generating protobufs using the
`--go_out=plugins=grpc` flag, and we don't get the new exported variable
name. We want the exported name because we intend to place the generated
*.orion.pb.go files in a separate package. Currently because the
generated *.orion.pb.go file depends on an unexported variable, it _has_
to be placed in the same package as the generated protobufs.